requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
black==24.8.0
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
distlib==0.3.8
distro==1.9.0
Django==5.0.7
djangorestframework==3.15.2
drf-yasg==1.21.7
dulwich==0.21.7
fastjsonschema==2.19.1
filelock==3.13.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.3.0
keyring==24.3.0
more-itertools==10.2.0
msgpack==1.0.7
mypy-extensions==1.0.0
openai==1.30.3
packaging==23.2
pathspec==0.12.1
pexpect==4.9.0
Pillow==6.2.2
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
ptyprocess==0.7.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pyproject_hooks==1.0.0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.6.1
requests==2.31.0
requests-toolbelt==1.0.0
sentry-sdk==1.40.0
shellingham==1.5.4
sniffio==1.3.1
sqlparse==0.5.1
tomlkit==0.12.3
tqdm==4.66.4
trove-classifiers==2024.1.8
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.0
uvicorn==0.27.0.post1
virtualenv==20.25.0
xattr==0.10.1
zipp==3.17.0
